Snowflake : progress after bug fix
Fixed stupid bug with WorldCameraPos beaing a Vec3 only, and thus not taking into account the world offset during matrix multiply.
Right now stereoscopy works, including with the old variables.

Added more refactoring and insights.
This is not raymarching, but finely tuned noise sampling and walk.
Lots of values are critical to the effect.

Tried using ddx/ddy for normal computation ; not great not bad, works but low resolution.
Next : try to use it with geometry shader...