Portal hooks system
Experiments with detecting and modifying VRC Portals.
I can detect new portals and manipulate them (relocation, deletion, etc).
The world texture and text info can be extracted, so the graphics can be completely changed if needed.
Sadly I cannot use the info to change the lifetime / save it as a static portal.