Fix spotlight shader bugs
Garvas had problems configuring it.
After debugging, it seems that the shader mode changes a lot depending on the config :
- Light colors will use emissive with premultiplied alpha, and no discard.
- Black colors must use alpha blend and discard.
My config used black, so this behavior was not obvious.
Now just apply reveal alpha to all of final color to be sure.

Also added blend for end of range.