Spatial HUD : final fixes
Doc, spacing.

Sadly it suffers from numerical instability at large world coordinates in its current use.
However this is not the shaders fault, but its use as a window close to the camera that amplifies the problem.
Experiments with floating point error cancellation did not improve anything, as the respective object and camera positions are already low resolution when reaching the shader.