Overlay upgrades : fullscreen, depth revisit, gamma range
Overlays are starting to be used on the glasses of lucifer.
- Add a fullscreen screenspace mode to prevent near-plane clipping of the overlay for : GammaAdjust, Normals and Wireframe.
- Enlarge gamma range for more flexibility ; restrictions of the range are reserved for animators.

Fullscreen required using a geometry pass for proper screenspace positionning.

Revisit of depth code.
Tried improving it using bgolus nice writeup and examples : https://gist.github.com/bgolus/a07ed65602c009d5e2f753826e8078a0.
bgolus code is nice as it only uses the pixel pos to infer view/world positions.
This would fit nicely with the fullscreen mode that overrides clip space position for the screenspace effect.

Testing bgolus code in VR generates mismatches between eyes.
Lots of tests followed, here are the notes :
We do not have access to the inverse projection matrix (CameraInvProjection is not inv P !).
Current depth shader can bypass it by relying on existing world positions and matching their interpolations with screenspace UVs and the depthTexture.
Sadly using the fullscreen screenspace mode does not give us world positions, as we synthetise clip space directly.
Inverting the P matrix in shader is expensive and unprecise.
Using the world clip planes also requires a matrix inversion (solve 3 plane intersection), and it is not sure if they are correct for oblique matrices (VR).
The current strategy in normals/wireframe uses CameraInvProjection, but this is not true VS in my experiments, and only works for normal-like effects, but not position.
Thus FrameGrid cannot be converted, and stays with its old depth code.
I am not even sure what the depth code actually does in normals/wireframe...