Heavy revisit of PN_Quads, added better bezier patch one
PN quads is cheap-ish but only C1 on quad edges.
Thus using the TBN from dPN(u,v) for vertex displacement leads to tears on edges.
This is bad for procedural ruffles, and cannot be fixed without making the TBN smoothly from another strategy (and mismatch from the PN surface).

Started a strategy with bezier patch that are made smoother, and should hopefully reach C2 on the edges.
TODO make the TBN for them.

For both shaders, added toggle to show TBN for smoothed curve.
Nice for debug and understanding.