Bezier patch TBN with derivatives : not C2 on edges
Added derivatives computation to generate TBN on interpolated surface.
Sadly still not C2.
The only solution seems to do a separate interpolation (quadratic) for TBN.

Another comparison shows that the surface of bezier is not that different from PN, and maybe not worth the extra cost.
Most of the differences probably come from definitions of the displacement edge vectors.