PN Quads : tested normal interpolation from paper for C1 TBN : failure
This interpolation is only described for the normal field.
Tangent and binormal have been improvised : use interpolation on the vector which is perpendicular to the edge like the normal, and cross for the second one.
This does not work well in practice.

Computing the center control points should be done in the hull constant function, but triggers DX11 shader compiler "IPC" errors.
Also the approximation is quite bad even for normal.
Thus this is a dead end, save if for posterity.

Next idea : use derivatives but twaerk them to be C1 near the edges.