Procedural ruffles tearing : fixed !
Used my C1 bezier quad as the base interpolated surface for procedural ruffles.
This fixes the tearing issues that were due to non C1 TBN on the edges.
TBN was used to guide ruffle splines, so it was required to find a way to make it continuous.

Implementation of procedural ruffles still needs cleanup, some checks to make it robust, and integration in orels for a real test.